--- backend/internal/entities/services/service_helper.go
@@ -1,16 +1,0 @@
-package services
-
-import (
-	"github.com/gin-gonic/gin"
-)
-
-type ErrorResponse struct {
-	Message string `json:"message"`
-}
-
-func handleError(ctx *gin.Context, status int, err error) {
-	res := ErrorResponse{
-		Message: err.Error(),
-	}
-	ctx.JSON(status, &res)
-}
